refactor: create v1.api.WeaviatClient in tests directly
