+++ src/it/java/io/weaviate/integration/DataITest.java
@@ -18,7 +18,7 @@ import io.weaviate.client6.v1.api.collections.vectorizers.NoneVectorizer;
 import io.weaviate.containers.Container;
 
 public class DataITest extends ConcurrentTest {
-  private static WeaviateClient client = Container.WEAVIATE.getClient().apiClient();
+  private static WeaviateClient client = Container.WEAVIATE.getClient();
   private static final String COLLECTION = unique("Artists");
   private static final String VECTOR_INDEX = "bring_your_own";
